* move engineNames to engine-names * move setup and get engine calls to unique functions (see namespace) * split up hooks into separate files * create useUrlParams hook * started working-key tests
