+++ src/hooks/index.js
@@ -1,1 +1,7 @@
-module.exports = require('./hooks')
+const useEffect = require('./useEffect')
+const useState = require('./useState')
+const useStore = require('./useStore')
+const useUrlParams = require('./useUrlParams')
+module.exports = {
+  useEffect, useState, useStore, useUrlParams
+}
